Выполнены доработки по замечаниям после ревью
